# penguins-incus-platform/api/schema/openapi.yaml
openapi: "3.1.0"
info:
  title: Penguins-Incus-Platform API
  version: "0.1.0"
  description: >
    Canonical API definition for the PIP daemon.
    All three frontends (QML, web, CLI) are thin clients against this API.
    The D-Bus interface mirrors every operation defined here.
  license:
    name: GPL-3.0-or-later
servers:
  - url: http://127.0.0.1:8765
    description: Local PIP daemon
...
components:
  schemas:
    Error:
      type: object
      required: [code, message]
      properties:
        code:    { type: integer }
        message: { type: string }
...
    Remote:
      type: object
      required: [name, url]
      properties:
        name:     { type: string }
        url:      { type: string, format: uri }
        protocol:
          type: string
          enum: [incus, lxd, simplestreams]
          default: incus
        auth_type:
          type: string
          enum: [tls, oidc]
          default: tls
        public:   { type: boolean, default: false }
...
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Conflict:
      description: Resource already exists or state conflict
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
paths:
...
  /api/v1/remotes:
    get:
      operationId: listRemotes
      summary: List configured remote servers
      tags: [remotes]
      responses:
        "200":
          description: Remote list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Remote" }
    post:
      operationId: addRemote
      summary: Add a remote server
      tags: [remotes]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Remote" }
      responses:
        "201":
          description: Remote added
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/remotes/{name}:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string }
    get:
      operationId: getRemote
      summary: Get remote details
      tags: [remotes]
      responses:
        "200":
          description: Remote detail
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Remote" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: removeRemote
      summary: Remove a remote server
      tags: [remotes]
      responses:
        "204":
          description: Removed
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/remotes/{name}/activate:
    parameters:
      - name: name
        in: path
        required: true
        schema: { type: string }
    put:
      operationId: activateRemote
      summary: Switch the active remote for all subsequent API calls
      tags: [remotes]
      responses:
        "200":
          description: Active remote switched
          content:
            application/json:
              schema:
                type: object
                properties:
                  active: { type: string }
        "404": { $ref: "#/components/responses/NotFound" }

  # ── Operations ─────────────────────────────────────────────────────────────
...